+++ src/musort/tools/__init__.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from .clargparser import clargs
 from .os_locale import REPLACEMENTS, is_hidden
 
-__all__ = ["clargs", "REPLACEMENTS", "is_hidden", "is_ok", "iterdir", "cache", "cleanup"]
+__all__ = ["clargs", "REPLACEMENTS", "is_hidden", "is_ok", "iterdir", "cleanup"]
 
 
 def is_ok(p: Path, /):
@@ -23,22 +23,6 @@ def iterdir(dir: Path, /) -> Iterable[Path]:
     for p in dir.iterdir():
         if is_ok(p):
             yield p
-
-
-class cache:
-    _genre: dict[str, str | None] = {}
-    """Mapping of artist name to genre. Unused if --single-genre is disabled."""
-
-    @classmethod
-    def genre(cls, artist: str, default: str | None):
-        """Retrieve an item from the cache."""
-        if not clargs.single_genre:
-            return default
-        artist = artist.lower()
-        if artist in cls._genre:
-            return cls._genre[artist]
-        cls._genre[artist] = default
-        return default
 
 
 def cleanup(dirs: Iterable[Path]):